Routing integration, security restrictions, loading messages.
Made it so everything routes properly so we can expand out 'list' feature into a corresponding list/details.
Built the main navigation components and routes for the app.  Added an appointments list piece.
Forced everything to go through login if the user is not currently authenticated.
Created a login component and centralized the aws amplify auth piece there.
Added loading messages on the data for the UI so we don't get screen flashes of content.