Database API endpoints for individual entities.
Made it so you can retrieve the individual entities for the Client,Therapist, & Appointment stuff.
Made it so selecting these entities get logged to the ActivityLog.
Updated the database schema to handle the ActivityLog.

Refactored the api file into individual endpoint files to make it easier to isolate and extend the code.