Linked amazon cognito session w/ server api session.
Made it so new users that sign up are added as system users.  Made it so the login form is styled.
Made it so the server API can only be used by authenticated users.  Added logout component and fixed a number of state
issues with the amazon cognito and server side state.