Final UI changes, login sync issues.
Made it so if the PHP server times out it logs the user out of Amazon and sends them to the login page (based on the 401 header status).

Made some UI adjustments.

Made it so adding an appointment for the client or the therapist preselects them on the schedule page.